feat: introduce quantum-enhanced performance analytics module with consciousness integration and temporal analysis capabilities

+++ engine/oxirs-vec/src/content_processing/office_handlers.rs
@@ -205,7 +205,7 @@ impl DocxHandler {
                                 page: None,
                                 section: None,
                                 char_offset: None,
-                                line: Some(i as u32 + 1),
+                                line: Some(i + 1),
                                 column: None,
                             },
                         });
